+++ project_scripts/analysis/analysis_groups.py
@@ -104,6 +104,7 @@ ASSET_GROUPS = {
         "name": "Government Bond Yields",
         "description": "US, Germany, UK and Japan yields.",
         "assets": [
+            "US3M",
             "US2Y",
             "US10Y",
             "US30Y",
@@ -115,8 +116,9 @@ ASSET_GROUPS = {
 
     "us_yield_curve": {
         "name": "US Yield Curve",
-        "description": "US 2Y, 10Y and 30Y yields.",
+        "description": "US 3M, 2Y, 10Y and 30Y yields.",
         "assets": [
+            "US3M",
             "US2Y",
             "US10Y",
             "US30Y"